requirements-mac.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements-mac.txt
absl-py==2.3.1
    # via tensorboard
accelerate==1.9.0
    # via
    #   sentence-transformers
    #   transformers
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.14
    # via fsspec
aiosignal==1.4.0
    # via aiohttp
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   google-genai
    #   httpx
    #   openai
appnope==0.1.4
    # via ipykernel
asttokens==3.0.0
    # via stack-data
attrs==25.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==5.5.2
    # via
    #   google-auth
    #   streamlit
certifi==2025.7.14
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via streamlit
comm==0.2.2
    # via
    #   ipykernel
    #   ipywidgets
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
datasets==4.0.0
    # via sentence-transformers
debugpy==1.8.15
    # via ipykernel
decorator==5.2.1
    # via ipython
dill==0.3.8
    # via
    #   datasets
    #   multiprocess
distro==1.9.0
    # via openai
dotenv==0.9.9
    # via -r requirements.in
executing==2.2.0
    # via stack-data
filelock==3.18.0
    # via
    #   datasets
    #   huggingface-hub
    #   torch
    #   transformers
fonttools==4.59.0
    # via matplotlib
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.3.0
    # via
    #   datasets
    #   huggingface-hub
    #   torch
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
google-auth==2.40.3
    # via google-genai
google-genai==1.26.0
    # via -r requirements.in
grpcio==1.73.1
    # via tensorboard
h11==0.16.0
    # via httpcore
hf-xet==1.1.5
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   google-genai
    #   openai
huggingface-hub==0.33.4
    # via
    #   accelerate
    #   datasets
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
ipykernel==6.29.5
    # via -r requirements.in
ipython==9.4.0
    # via
    #   ipykernel
    #   ipywidgets
ipython-pygments-lexers==1.1.1
    # via ipython
ipywidgets==8.1.7
    # via -r requirements.in
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
    #   torch
jiter==0.10.0
    # via openai
joblib==1.5.1
    # via scikit-learn
jsonschema==4.24.1
    # via altair
jsonschema-specifications==2025.4.1
    # via jsonschema
jupyter-client==8.6.3
    # via ipykernel
jupyter-core==5.8.1
    # via
    #   ipykernel
    #   jupyter-client
jupyterlab-widgets==3.0.15
    # via ipywidgets
kiwisolver==1.4.8
    # via matplotlib
markdown==3.8.2
    # via tensorboard
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
matplotlib==3.10.3
    # via -r requirements.in
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
mpmath==1.3.0
    # via sympy
multidict==6.6.3
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.16
    # via datasets
narwhals==1.47.0
    # via altair
nest-asyncio==1.6.0
    # via ipykernel
networkx==3.5
    # via torch
numpy==2.3.1
    # via
    #   -r requirements.in
    #   accelerate
    #   contourpy
    #   datasets
    #   matplotlib
    #   pandas
    #   pydeck
    #   scikit-learn
    #   scipy
    #   streamlit
    #   tensorboard
    #   transformers
openai==1.97.0
    # via -r requirements.in
packaging==25.0
    # via
    #   accelerate
    #   altair
    #   datasets
    #   huggingface-hub
    #   ipykernel
    #   matplotlib
    #   streamlit
    #   tensorboard
    #   transformers
pandas==2.3.1
    # via
    #   -r requirements.in
    #   datasets
    #   streamlit
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==11.3.0
    # via
    #   matplotlib
    #   sentence-transformers
    #   streamlit
platformdirs==4.3.8
    # via jupyter-core
prompt-toolkit==3.0.51
    # via ipython
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
protobuf==6.31.1
    # via
    #   streamlit
    #   tensorboard
psutil==7.0.0
    # via
    #   accelerate
    #   ipykernel
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pyarrow==20.0.0
    # via
    #   datasets
    #   streamlit
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.7
    # via
    #   google-genai
    #   openai
pydantic-core==2.33.2
    # via pydantic
pydeck==0.9.1
    # via streamlit
pygments==2.19.2
    # via
    #   ipython
    #   ipython-pygments-lexers
pyparsing==3.2.3
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   jupyter-client
    #   matplotlib
    #   pandas
python-dotenv==1.1.1
    # via dotenv
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   accelerate
    #   datasets
    #   huggingface-hub
    #   transformers
pyzmq==27.0.0
    # via
    #   ipykernel
    #   jupyter-client
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via transformers
requests==2.32.4
    # via
    #   datasets
    #   google-genai
    #   huggingface-hub
    #   streamlit
    #   transformers
rpds-py==0.26.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
safetensors==0.5.3
    # via
    #   accelerate
    #   transformers
scikit-learn==1.7.0
    # via
    #   -r requirements.in
    #   sentence-transformers
scipy==1.16.0
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==5.0.0
    # via -r requirements.in
setuptools==80.9.0
    # via
    #   tensorboard
    #   torch
six==1.17.0
    # via
    #   python-dateutil
    #   tensorboard
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   openai
stack-data==0.6.3
    # via ipython
streamlit==1.47.0
    # via -r requirements.in
sympy==1.14.0
    # via torch
tenacity==8.5.0
    # via
    #   google-genai
    #   streamlit
tensorboard==2.19.0
    # via -r requirements.in
tensorboard-data-server==0.7.2
    # via tensorboard
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.21.2
    # via transformers
toml==0.10.2
    # via streamlit
torch==2.7.1
    # via
    #   -r requirements.in
    #   accelerate
    #   sentence-transformers
    #   transformers
tornado==6.5.1
    # via
    #   ipykernel
    #   jupyter-client
    #   streamlit
tqdm==4.67.1
    # via
    #   -r requirements.in
    #   datasets
    #   huggingface-hub
    #   openai
    #   sentence-transformers
    #   transformers
traitlets==5.14.3
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
transformers==4.53.2
    # via
    #   -r requirements.in
    #   sentence-transformers
typing-extensions==4.14.1
    # via
    #   aiosignal
    #   altair
    #   anyio
    #   google-genai
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sentence-transformers
    #   streamlit
    #   torch
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
watchdog==6.0.0
    # via -r requirements.in
wcwidth==0.2.13
    # via prompt-toolkit
websockets==15.0.1
    # via google-genai
werkzeug==3.1.3
    # via tensorboard
widgetsnbextension==4.0.14
    # via ipywidgets
xxhash==3.5.0
    # via datasets
yarl==1.20.1
    # via aiohttp
